pyproject.toml
[tool.poetry]
name = "catalogue"
version = "0.1.0"
description = ""
authors = ["Andrew Phelps <andrew.m.phelps@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {version = "^0.194.4", extras = ["debug-server"]}
sqlalchemy = "^2.0.18"
fastapi = "^0.100.0"
toolz = "^0.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

